feat(select): Change root to inline-block & add fullwidth flag
This accommodates upcoming change which will allow the menu to automagically match width to the select anchor. As a side fix, the anchor element has been changed from `inline-flex` -> `flex` since it should always respect the width of the root element.

Expect the next change to have menu auto-size its width to match select by default, thus rendering any custom sizing `demo-width-class` on the MENU to be obsolete.

BREAKING CHANGE: root of mdc-select is now an inline-block element, use custom width class (i.e. `demo-width-class`) on the root instead of the anchor for width adjustments; alternately, use the new `mdc-select--fullwidth` on the root to expand width to that of its parent container

PiperOrigin-RevId: 307907276